refactor: move dashboard to root route and login to /login route
- Move dashboard page from /dashboard to root route (/)
- Move login page from root route to /login route
- Update middleware to protect root route and redirect to /login
- Update all route references in components (page-header, dev-navigation)
- Remove /dashboard route
