+++ packages/tv-ad-admin-next/app/page.tsx
@@ -1,267 +1,182 @@
 'use client'
 
-import { useState, useEffect } from 'react'
+import { useCallback, useEffect, useRef, useState } from 'react'
 
-import { useRouter } from 'next/navigation'
+import Link from 'next/link'
 
-import type { SendOtpResponse } from '@/types/api'
-
-import EmailForm from '@/components/login/email-form'
-import OptForm from '@/components/login/opt-form'
+import StateCard from '@/components/dashboard/state-card'
 import PageHeader from '@/components/shared/page-header'
 import PageMain from '@/components/shared/page-main'
-import { AUTH_MESSAGES, LOADING_MESSAGES } from '@/constants/messages'
-import { useAuthStore } from '@/store'
-import { validateEmail } from '@/utils/validation'
+import { Button } from '@/components/ui/button'
+import {
+  Card,
+  CardHeader,
+  CardTitle,
+  CardDescription,
+  CardContent,
+} from '@/components/ui/card'
+import { OrderStateMap } from '@/constants'
+import FileIcon from '@/public/icons/file.svg'
+import UploadIcon from '@/public/icons/upload.svg'
+import { getOrdersState } from '@/utils/order-grouping'
+
+const ERROR_MESSAGE = '載入訂單失敗，請稍後再試'
 
 export default function HomePage() {
-  const router = useRouter()
-  const { isAuthenticated, login, initialize } = useAuthStore()
+  const [ordersState, setOrdersState] = useState<
+    { state: keyof typeof OrderStateMap; count: number }[]
+  >([])
+  const [error, setError] = useState<string | null>(null)
+  const [isLoading, setIsLoading] = useState(true)
+  const abortControllerRef = useRef<AbortController | null>(null)
 
-  const [showOtpForm, setShowOtpForm] = useState(false)
-  const [email, setEmail] = useState('')
-  const [isLoading, setIsLoading] = useState(false)
-  const [loadingMessage, setLoadingMessage] = useState('')
-  const [error, setError] = useState('')
-  const [countdown, setCountdown] = useState(0)
-  const [canResend, setCanResend] = useState(true)
-  const [failedAttempts, setFailedAttempts] = useState(0)
-  const maxAttempts = 3 // 與後端一致
-
-  // 初始化認證狀態
-  useEffect(() => {
-    initialize()
-  }, [initialize])
-
-  // 如果已登入，重定向到 dashboard
-  useEffect(() => {
-    if (isAuthenticated) {
-      router.push('/dashboard')
+  const fetchOrders = useCallback(async () => {
+    // 取消之前的請求
+    if (abortControllerRef.current) {
+      abortControllerRef.current.abort()
     }
-  }, [isAuthenticated, router])
 
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault()
-    setError('')
+    abortControllerRef.current = new AbortController()
     setIsLoading(true)
-    setLoadingMessage(LOADING_MESSAGES.CHECKING_MEMBER)
+    setError(null)
 
     try {
-      // 驗證輸入
-      const validation = validateEmail(email)
-      if (!validation.isValid) {
-        setError(validation.error || AUTH_MESSAGES.EMAIL_INVALID)
-        return
-      }
-
-      // 呼叫 Next.js API Route 發送 OTP
-      const response = await fetch('/api/auth/send-otp', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          email,
-        }),
+      const res = await fetch('/api/dashboard/state', {
+        signal: abortControllerRef.current.signal,
       })
 
-      const data: SendOtpResponse = await response.json()
-
-      if (!data.success) {
-        setError(data.message || AUTH_MESSAGES.SEND_OTP_FAILED)
-        return
+      // 先讀取 response text，以便在 JSON 解析失敗時可以記錄
+      let data
+      let responseText: string | null = null
+      try {
+        responseText = await res.text()
+        data = JSON.parse(responseText)
+      } catch (jsonError) {
+        // JSON 解析失敗（可能是伺服器回傳非 JSON，如 HTML 錯誤頁）
+        console.error('[Dashboard Stats API] JSON 解析失敗:', {
+          status: res.status,
+          statusText: res.statusText,
+          url: res.url,
+          responseText: responseText?.substring(0, 500), // 限制長度避免 log 過長
+          error: jsonError,
+        })
+        throw new Error(ERROR_MESSAGE)
       }
 
-      // 會員驗證通過，開始發送 OTP
-      setLoadingMessage(LOADING_MESSAGES.SENDING_OTP)
-
-      // 開發環境：在瀏覽器 Console 顯示驗證碼（帶顏色）
-      if (data.data?.otp) {
-        console.log(
-          '%c🔐 ========== OTP 驗證碼 ==========',
-          'color: #10b981; font-size: 14px; font-weight: bold;'
-        )
-        console.log(`%c📧 ${email}`, 'color: #3b82f6; font-size: 12px;')
-        console.log(
-          `%c🔢 驗證碼: ${data.data.otp}`,
-          'color: #ef4444; font-size: 16px; font-weight: bold;'
-        )
-        console.log('%c⏰ 有效期: 5 分鐘', 'color: #f59e0b; font-size: 12px;')
-        console.log(
-          '%c=====================================',
-          'color: #10b981; font-size: 14px;'
-        )
+      if (!res.ok) {
+        // API 回傳了錯誤 JSON，記錄完整的 debug 資訊
+        console.error('[Dashboard Stats API] 請求失敗:', {
+          status: res.status,
+          statusText: res.statusText,
+          url: res.url,
+          apiError: data.error,
+          fullResponse: data,
+        })
+        throw new Error(ERROR_MESSAGE)
       }
 
-      setShowOtpForm(true)
-      setCountdown(60)
-      setCanResend(false)
-    } catch (err) {
-      console.error(err)
-      setError(AUTH_MESSAGES.NETWORK_ERROR)
-    } finally {
-      setIsLoading(false)
-      setLoadingMessage('')
-    }
-  }
-
-  // 倒數計時效果
-  useEffect(() => {
-    let timer: ReturnType<typeof setTimeout>
-    if (countdown > 0) {
-      timer = setTimeout(() => {
-        setCountdown(countdown - 1)
-      }, 1000)
-    } else {
-      setCanResend(true)
-    }
-    return () => clearTimeout(timer)
-  }, [countdown])
-
-  // 驗證碼驗證
-  const handleOtpSubmit = async (value?: string) => {
-    setError('')
-    setIsLoading(true)
-
-    try {
-      // 驗證 OTP 格式
-      const otpValue = value
-      if (!otpValue?.trim()) {
-        setError(AUTH_MESSAGES.OTP_REQUIRED)
-        return
+      // 成功：設定資料
+      setOrdersState(getOrdersState(data.orders || []))
+    } catch (error) {
+      if (error instanceof Error && error.name === 'AbortError') {
+        return // 請求被取消，不更新狀態
       }
-
-      // 呼叫 Next.js API Route 驗證 OTP
-      const response = await fetch('/api/auth/verify-otp', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        credentials: 'include', // 重要：包含 cookies
-        body: JSON.stringify({
-          email,
-          otp: otpValue,
-        }),
+      // 記錄完整的錯誤資訊供工程師 debug
+      console.error('[Dashboard Stats] 載入訂單失敗:', {
+        error,
+        errorMessage: error instanceof Error ? error.message : String(error),
+        errorStack: error instanceof Error ? error.stack : undefined,
+        errorName: error instanceof Error ? error.name : undefined,
       })
-
-      const data = await response.json()
-
-      if (!data.success) {
-        const newFailedAttempts = failedAttempts + 1
-        setFailedAttempts(newFailedAttempts)
-
-        if (newFailedAttempts >= maxAttempts) {
-          setError(AUTH_MESSAGES.OTP_TOO_MANY_ATTEMPTS)
-        } else {
-          setError(data.message || AUTH_MESSAGES.OTP_INCORRECT)
-        }
-        return
-      }
-
-      // 登入成功，更新 store
-      if (data.user) {
-        // login 是同步函數，只是更新 zustand store 的 state，不需要 await
-        login(data.user)
-      }
-
-      // 登入成功，跳轉到 dashboard
-      // Cookie 已在 API 響應時設定，狀態更新是同步的，可直接跳轉
-      router.push('/dashboard')
-    } catch (err) {
-      console.error(err)
-      setError(AUTH_MESSAGES.NETWORK_ERROR)
+      // 只顯示友善的錯誤訊息給使用者，避免顯示技術細節
+      // 技術細節已記錄在 console，供工程師 debug 使用
+      setError(ERROR_MESSAGE)
     } finally {
       setIsLoading(false)
     }
-  }
+  }, [])
 
-  const handleResendOtp = async () => {
-    if (!canResend) return
-
-    setError('')
-    setFailedAttempts(0) // 重置失敗次數
-    setIsLoading(true)
-    setLoadingMessage(LOADING_MESSAGES.CHECKING_MEMBER)
-
-    try {
-      // 呼叫 Next.js API Route 重新發送 OTP
-      const response = await fetch('/api/auth/send-otp', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          email,
-        }),
-      })
-
-      const data: SendOtpResponse = await response.json()
-
-      if (!data.success) {
-        setError(data.message || AUTH_MESSAGES.RESEND_OTP_FAILED)
-        return
+  useEffect(() => {
+    fetchOrders()
+    return () => {
+      if (abortControllerRef.current) {
+        abortControllerRef.current.abort()
       }
-
-      // 會員驗證通過，開始發送 OTP
-      setLoadingMessage(LOADING_MESSAGES.SENDING_OTP)
-
-      // 開發環境：在瀏覽器 Console 顯示驗證碼（帶顏色）
-      if (data.data?.otp) {
-        console.log(
-          '%c🔐 ========== 重新發送 OTP ==========',
-          'color: #10b981; font-size: 14px; font-weight: bold;'
-        )
-        console.log(`%c📧 ${email}`, 'color: #3b82f6; font-size: 12px;')
-        console.log(
-          `%c🔢 驗證碼: ${data.data.otp}`,
-          'color: #ef4444; font-size: 16px; font-weight: bold;'
-        )
-        console.log('%c⏰ 有效期: 5 分鐘', 'color: #f59e0b; font-size: 12px;')
-        console.log(
-          '%c=====================================',
-          'color: #10b981; font-size: 14px;'
-        )
-      }
-
-      setCountdown(60)
-      setCanResend(false)
-    } catch (err) {
-      console.error(err)
-      setError(AUTH_MESSAGES.NETWORK_ERROR)
-    } finally {
-      setIsLoading(false)
-      setLoadingMessage('')
     }
-  }
+  }, [fetchOrders])
 
   return (
     <>
-      <PageHeader variant="centered" />
-      <PageMain className="flex justify-center py-5 md:py-10">
-        <div className="flex h-fit max-w-[288px] flex-col items-center rounded-xl border border-border-default bg-surface-primary p-4 shadow-lg md:max-w-[448px] md:min-w-[448px] md:p-6">
-          {!showOtpForm ? (
-            <EmailForm
-              email={email}
-              setEmail={setEmail}
-              handleSubmit={handleSubmit}
-              isLoading={isLoading}
-              loadingMessage={loadingMessage}
-              error={error}
-            />
-          ) : (
-            <OptForm
-              email={email}
-              error={error}
-              isLoading={isLoading}
-              countdown={countdown}
-              canResend={canResend}
-              failedAttempts={failedAttempts}
-              maxAttempts={maxAttempts}
-              handleOtpSubmit={handleOtpSubmit}
-              handleResendOtp={handleResendOtp}
-            />
-          )}
+      <PageHeader variant="spread" title="鏡新聞個人廣告系統" />
+      <PageMain className="grid grid-rows-[auto_1fr] gap-4 py-5 md:gap-10 md:py-10">
+        {/* --- Top two cards --- */}
+        <div className="grid grid-cols-1 gap-4 md:grid-cols-2 md:gap-6">
+          {/* Upload Card */}
+          <Link href="/upload">
+            <Card className="cursor-pointer items-center justify-center gap-3 hover:shadow-[0_4px_8px_0_rgba(0,0,0,0.10)]">
+              <UploadIcon className="size-10 text-blue-7" />
+              <CardTitle className="flex flex-col items-center gap-1">
+                <span>上傳廣告素材</span>
+                <CardDescription>上傳後即可進入製作流程</CardDescription>
+              </CardTitle>
+            </Card>
+          </Link>
+
+          {/* history Card */}
+          <Link href="/list">
+            <Card className="cursor-pointer items-center justify-center gap-3 hover:shadow-[0_4px_8px_0_rgba(0,0,0,0.10)]">
+              <FileIcon className="size-10 text-blue-7" />
+              <CardTitle className="flex flex-col items-center gap-1">
+                <span>訂單紀錄</span>
+                <CardDescription>查看與管理所有訂單</CardDescription>
+              </CardTitle>
+            </Card>
+          </Link>
         </div>
+
+        {/* --- Bottom: Order state overview --- */}
+        <Card>
+          <CardHeader>
+            <CardTitle>訂單狀態總覽</CardTitle>
+          </CardHeader>
+
+          <CardContent className="grid grid-cols-1 gap-2 md:grid-cols-4 md:gap-4 xl:grid-cols-6">
+            {isLoading ? (
+              <div className="col-span-full flex items-center justify-center py-8">
+                <div className="flex flex-col items-center gap-3">
+                  <div className="size-8 animate-spin rounded-full border-4 border-gray-3 border-t-blue-6" />
+                  <p className="text-sm text-gray-7">載入中...</p>
+                </div>
+              </div>
+            ) : error ? (
+              <div className="col-span-full flex flex-col items-center justify-center gap-4 py-8">
+                <p className="text-sm text-red-9">{error}</p>
+                <Button onClick={fetchOrders}>重新整理</Button>
+              </div>
+            ) : ordersState.length === 0 ? (
+              <div className="col-span-full flex items-center justify-center py-8">
+                <p className="text-sm text-gray-7">目前沒有訂單狀態資料</p>
+              </div>
+            ) : (
+              ordersState.map(({ state, count }) => {
+                const config =
+                  OrderStateMap[state as keyof typeof OrderStateMap]
+                if (!config) return null
+
+                return (
+                  <Link key={state} href={`/list?state=${state}`}>
+                    <StateCard
+                      count={count}
+                      text={config.label}
+                      color={config.colors.text}
+                      bgColor={config.colors.bg}
+                    />
+                  </Link>
+                )
+              })
+            )}
+          </CardContent>
+        </Card>
       </PageMain>
     </>
   )
