+++ packages/tv-ad-admin-next/components/demo/dev-navigation.tsx
@@ -24,10 +24,10 @@ export default function DevNavigation() {
           Demo
         </Link>
         <span className="font-extrabold">|</span>
-        <Link href="/" className="hover:text-purple-400">
+        <Link href="/login" className="hover:text-purple-400">
           Login
         </Link>
-        <Link href="/dashboard" className="hover:text-indigo-400">
+        <Link href="/" className="hover:text-indigo-400">
           Dashboard
         </Link>
         <Link href="/upload" className="hover:text-pink-400">
